Online-Training-Course-in-ASP.NET-MVC
This is a web-based project. I have followed the Agile model on (SDLC) to build this project. There are three types of user in this web-based application (Student, Teacher, Admin) On this website, a student can register for a course they desire. A course ID will be provided for every course. Student can register through the system. An id will generate automatically for a student. The information of the student will be stored in the system database.

A teacher can also register as a teacher into the system. A teacher ID will also generate automatically for the teacher. The teacher can upload course video into the system. The teacher will be able to take the exam from his/her specific courses.

A profile page will be provided to both teachers and the students. Students can register the courses with their student ID and Password. The registered student will able to watch which courses have been taken by them from their profile.

There will be an admin to control the whole system. The admin will also have a login system. A profile will also be provided to the admin. An admin can be able to monitor student information and teacher information and which courses are taken by which students and which courses are teaching by which teachers. Admin can also approve a teacher when a teacher gets registered. Admin can also approve the course videos which will be uploaded by the teachers.

I have used Visual Studio as IDE and used SQL sever to build the database of this project.